feat: refactor organization and user resources, add scheduled task management
